# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduSaaS - Plataforma Educacional Multi-tenant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List exams
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, active, closed]
        - name: type
          in: query
          schema:
            type: string
            enum: [enem, simulado, traditional, homework]
        - name: classId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "201":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{examId}:
    get:
      operationId: getExam
      tags: [exams]
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      summary: Get exam with questions
      responses:
        "200":
          description: Exam detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
    put:
      operationId: updateExam
      tags: [exams]
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /exams/{examId}/publish:
    post:
      operationId: publishExam
      tags: [exams]
      summary: Publish/schedule exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{examId}/report:
    get:
      operationId: getExamReport
      tags: [exams]
      summary: Get detailed exam report
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamReport"

  # ─── QUESTIONS ───────────────────────────────────────────────────────────────
  /exams/{examId}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      summary: List questions for an exam
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      summary: Create question in exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{questionId}:
    put:
      operationId: updateQuestion
      tags: [questions]
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── EXAM SESSIONS (student taking exam) ────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── AUTH ──────────────────────────────────────────────────────────────────
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [enem, simulado, traditional, homework]
        status:
          type: string
          enum: [draft, scheduled, active, closed]
        timeLimitMinutes:
          type: integer
          nullable: true
        startsAt:
          type: string
          nullable: true
        endsAt:
          type: string
          nullable: true
        maxAttempts:
          type: integer
          nullable: true
        classId:
          type: integer
          nullable: true
        subjectId:
          type: integer
          nullable: true
        isPublic:
          type: boolean
        showResultImmediately:
          type: boolean
        questionsCount:
          type: integer
        createdById:
          type: integer
        createdAt:
          type: string
      required: [id, tenantId, title, type, status, isPublic, showResultImmediately, questionsCount, createdById, createdAt]
    ExamDetail:
      allOf:
        - $ref: "#/components/schemas/Exam"
        - type: object
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
          required: [questions]
    CreateExamBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [enem, simulado, traditional, homework]
        timeLimitMinutes:
          type: integer
          nullable: true
        startsAt:
          type: string
          nullable: true
        endsAt:
          type: string
          nullable: true
        maxAttempts:
          type: integer
          nullable: true
        classId:
          type: integer
          nullable: true
        subjectId:
          type: integer
          nullable: true
        isPublic:
          type: boolean
          default: false
        showResultImmediately:
          type: boolean
          default: true
      required: [title, type]
    ExamReport:
      type: object
      properties:
        examId:
          type: integer
        title:
          type: string
        totalSessions:
          type: integer
        completedSessions:
          type: integer
        averageScore:
          type: number
          nullable: true
        averageTimeMinutes:
          type: number
          nullable: true
        questionStats:
          type: array
          items:
            $ref: "#/components/schemas/QuestionStat"
        scoreDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ScoreRange"
        topStudents:
          type: array
          items:
            $ref: "#/components/schemas/StudentRankItem"
      required: [examId, title, totalSessions, completedSessions, questionStats, scoreDistribution, topStudents]

    # ─── QUESTIONS ─────────────────────────────────────────────────────────────
    Question:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        type:
          type: string
          enum: [multiple_choice, true_false]
        statement:
          type: string
        explanation:
          type: string
          nullable: true
        topicId:
          type: integer
          nullable: true
        points:
          type: number
        order:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
      required: [id, examId, type, statement, points, order, options]
    QuestionOption:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        text:
          type: string
        isCorrect:
          type: boolean
        letter:
          type: string
      required: [id, questionId, text, isCorrect, letter]
    CreateQuestionBody:
      type: object
      properties:
        type:
          type: string
          enum: [multiple_choice, true_false]
        statement:
          type: string
        explanation:
          type: string
          nullable: true
        topicId:
          type: integer
          nullable: true
        points:
          type: number
          default: 1
        order:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/CreateOptionBody"
      required: [type, statement, order, options]
    CreateOptionBody:
      type: object
      properties:
        text:
          type: string
        isCorrect:
          type: boolean
        letter:
          type: string
      required: [text, isCorrect, letter]

    # ─── SESSIONS ──────────────────────────────────────────────────────────────
...
    QuestionStat:
      type: object
      properties:
        questionId:
          type: integer
        statement:
          type: string
        correctRate:
          type: number
        totalAnswers:
          type: integer
        optionDistribution:
          type: array
          items:
            $ref: "#/components/schemas/OptionCount"
      required: [questionId, statement, correctRate, totalAnswers, optionDistribution]
    OptionCount:
      type: object
      properties:
        optionId:
          type: integer
        letter:
          type: string
        count:
          type: integer
        isCorrect:
          type: boolean
      required: [optionId, letter, count, isCorrect]
    ScoreRange:
      type: object
      properties:
        range:
          type: string
        count:
          type: integer
      required: [range, count]
    StudentRankItem:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        score:
          type: number
        percentage:
          type: number
        rank:
          type: integer
      required: [userId, name, score, percentage, rank]